Version sent to Jason at Sunriver.
Fixed bug whereby if at Favorites page a user unfavoritizes an item, and that item was the last favorite in that category; and the user was previously viewing favorites at that page; then upon going back to that page, it would show no elements, and there wouldn't be a start.  Essentially the user could never view items in that category again!